New users

+++ api/admin_api.py
@@ -1,13 +1,13 @@
 from flask import abort, Blueprint, redirect, request
 from flask.templating import render_template
 
+from config.constants import ResearcherRole
 from config.settings import DOMAIN_NAME, DOWNLOADABLE_APK_URL, IS_STAGING
 from database.study_models import Study
-from database.user_models import Researcher
-from libs.admin_authentication import (
-    authenticate_system_admin, authenticate_admin_login, admin_is_system_admin,
-    get_admins_allowed_studies
-)
+from database.user_models import Researcher, StudyRelation
+from libs.admin_authentication import (assert_admin, assert_researcher_under_admin,
+    authenticate_researcher_login, authenticate_site_admin, get_researcher_allowed_studies,
+    get_session_researcher, researcher_is_an_admin)
 from libs.security import check_password_requirements
 
 admin_api = Blueprint('admin_api', __name__)
@@ -16,11 +16,14 @@ admin_api = Blueprint('admin_api', __name__)
 
 
 @admin_api.route('/add_researcher_to_study', methods=['POST'])
-@authenticate_system_admin
+@authenticate_site_admin
 def add_researcher_to_study():
     researcher_id = request.values['researcher_id']
     study_id = request.values['study_id']
-    Researcher.studies.through.objects.get_or_create(researcher_id=researcher_id, study_id=study_id)
+    assert_admin(study_id)
+    StudyRelation.objects.get_or_create(
+        study_id=study_id, researcher_id=researcher_id, relationship=ResearcherRole.researcher
+    )
 
     # This gets called by both edit_researcher and edit_study, so the POST request
     # must contain which URL it came from.
@@ -28,32 +31,43 @@ def add_researcher_to_study():
 
 
 @admin_api.route('/remove_researcher_from_study', methods=['POST'])
-@authenticate_system_admin
+@authenticate_site_admin
 def remove_researcher_from_study():
     researcher_id = request.values['researcher_id']
     study_id = request.values['study_id']
-    Researcher.objects.get(pk=researcher_id).studies.remove(study_id)
-
-    return redirect(request.values['redirect_url'])
-
-
-@admin_api.route('/delete_researcher/<string:researcher_id>', methods=['GET', 'POST'])
-@authenticate_system_admin
-def delete_researcher(researcher_id):
     try:
         researcher = Researcher.objects.get(pk=researcher_id)
     except Researcher.DoesNotExist:
         return abort(404)
-    
-    researcher.studies.clear()
+    assert_admin(study_id)
+    assert_researcher_under_admin(researcher)
+    StudyRelation.objects.filter(study_id=study_id, researcher_id=researcher_id).delete()
+    return redirect(request.values['redirect_url'])
+
+
+@admin_api.route('/delete_researcher/<string:researcher_id>', methods=['GET', 'POST'])
+@authenticate_site_admin
+def delete_researcher(researcher_id):
+    # only site admins can delete researchers from the system.
+    session_researcher = get_session_researcher()
+    if not session_researcher.site_admin:
+        return abort(403)
+
+    try:
+        researcher = Researcher.objects.get(pk=researcher_id)
+    except Researcher.DoesNotExist:
+        return abort(404)
+
+    StudyRelation.objects.filter(researcher=researcher).delete()
     researcher.delete()
     return redirect('/manage_researchers')
 
 
 @admin_api.route('/set_researcher_password', methods=['POST'])
-@authenticate_system_admin
+@authenticate_site_admin
 def set_researcher_password():
     researcher = Researcher.objects.get(pk=request.form.get('researcher_id', None))
+    assert_researcher_under_admin(researcher)
     new_password = request.form.get('password', '')
     if check_password_requirements(new_password, flash_message=True):
         researcher.set_password(new_password)
@@ -61,9 +75,10 @@ def set_researcher_password():
 
 
 @admin_api.route('/rename_study/<string:study_id>', methods=['POST'])
-@authenticate_system_admin
+@authenticate_site_admin
 def rename_study(study_id=None):
     study = Study.objects.get(pk=study_id)
+    assert_admin(study_id)
     new_study_name = request.form.get('new_study_name', '')
     study.name = new_study_name
     study.save()
@@ -74,12 +89,12 @@ def rename_study(study_id=None):
 
 
 @admin_api.route("/downloads")
-@authenticate_admin_login
+@authenticate_researcher_login
 def download_page():
     return render_template(
         "download_landing_page.html",
-        system_admin=admin_is_system_admin(),
-        allowed_studies=get_admins_allowed_studies(),
+        is_admin=researcher_is_an_admin(),
+        allowed_studies=get_researcher_allowed_studies(),
         domain_name=DOMAIN_NAME,
     )
 
@@ -90,25 +105,25 @@ def download_current():
 
 
 @admin_api.route("/download_debug")
-@authenticate_admin_login
+@authenticate_researcher_login
 def download_current_debug():
     return redirect("https://s3.amazonaws.com/beiwe-app-backups/release/Beiwe-debug.apk")
 
 
 @admin_api.route("/download_beta")
-@authenticate_admin_login
+@authenticate_researcher_login
 def download_beta():
     return redirect("https://s3.amazonaws.com/beiwe-app-backups/release/Beiwe.apk")
 
 
 @admin_api.route("/download_beta_debug")
-@authenticate_admin_login
+@authenticate_researcher_login
 def download_beta_debug():
     return redirect("https://s3.amazonaws.com/beiwe-app-backups/debug/Beiwe-debug.apk")
 
 
 @admin_api.route("/download_beta_release")
-@authenticate_admin_login
+@authenticate_researcher_login
 def download_beta_release():
     return redirect("https://s3.amazonaws.com/beiwe-app-backups/release/Beiwe-2.2.3-onnelaLabServer-release.apk")
